Move Generic Label interpolator code to ImageRegistration directory, add help output for all programs which can accept it
