+++ ImageRegistration/itkLabelImageGenericInterpolateImageFunction.h
