+++ ImageRegistration/itkLabelImageGenericInterpolateImageFunction.hxx
