+++ ImageRegistration/itkLabelSelectionAdaptor.h
